ai-chatbot/requirements.txt
flask==2.3.3
google-generativeai==0.3.1
python-dotenv==1.0.0
markdown==3.4.4
requests==2.31.0
markupsafe==2.1.3